webrtc-vlm-backend/requirements.txt
aioice==0.10.1
aiortc==1.13.0
av==14.4.0
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.6
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
google-crc32c==1.7.1
greenlet==3.2.4
h11==0.16.0
humanfriendly==10.0
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
pycparser==2.22
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.41
sympy==1.14.0
typing_extensions==4.14.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
